Added Pie Chart
Now pie chart is added to the page and legend is connected to both Pie
chart and parallel coordinates Tooltips won't follow the changes though
which is the next item to be fixed.
There is so much left there about colors and texts that I need to take
care of later but I leave them after the page is all together.