initial commit
This is the base for design explorer interface. I used bootstrap side
bar template and added a little bit of D3 on top of it. Now it works
pretty fine.  Not totally happy with the animation when it changes
between the layouts which I will address at some point later.